WOM-392: Update int arg types and try nodeid before read_browse_name()
Resolves issue with recursive scanning of PLC's
"Pointing.StaticPmParameters" tree not behaving the same as it does with
the CETC sim. The BrowseName of the PLC's nodes do not have [0]..[7] at
the end like the node ID, but SCU does not need to use the
BrowseName, as the name can be parsed from nodeid.Identifier
already. Doing this removes one asyncio call for each node during tree
scanning.

GUI also automatically releases command authority when disconnecting
from server.