WOM-405: Handle ConnectionError in SCU command functions
Added a ResultCode enum that contains the server's LastCmdResponse codes
from 0 to 11, but is extended with client side specific codes for
exceptions starting from -1 and decrementing.

A result code of -3 now signals to the Controller to handle a caught
closed connection exception by calling a new Model method that cleans
up the polling and event loop threads and signals to the MainView to
return to a disconnected state.