WOM-506: Do not block sending a 'TakeAuth' or 'ReleaseAuth' command
SCU's private 'user' and 'session_id' properties are now initialised to
UInt16(0) instead of None, to ensure they always are the correct type
for the 'TakeAuth' or 'ReleaseAuth' commands. And only the 'session_id'
is reset to UInt16(0) as a falsy value when any command responds with a
'NoCmdAuth' result code.